refactor: move sns packages
