+++ frontend/src/lib/components/sns-neuron-detail/SnsNeuronMaturityCard.svelte
