+++ frontend/src/lib/components/sns-neurons/SnsNeuronAge.svelte
