+++ frontend/src/lib/components/sns-neurons/SnsNeuronAmount.svelte
