+++ frontend/src/lib/components/sns-neurons/SnsNeuronCardTitle.svelte
@@ -29,8 +29,8 @@
 </div>
 
 <style lang="scss">
-  @use "@dfinity/gix-components/styles/mixins/media";
-  @use "@dfinity/gix-components/styles/mixins/card";
+  @use "../../../../node_modules/@dfinity/gix-components/styles/mixins/media";
+  @use "../../../../node_modules/@dfinity/gix-components/styles/mixins/card";
 
   .identifier {
     @include card.stacked-title;
