+++ frontend/src/lib/components/sns-neurons/SnsNeuronStateRemainingTime.svelte
